study_langgraph/pyproject.toml
[project]
name = "study-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.24",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.34",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
